requirements.txt
streamlit==1.4.0
Pillow==8.4.0
requests==2.26.0
streamlit-lottie==0.0.3
azure-storage-blob==12.9.0
altair==4.1.0
